@@ -1,45 +1,32 @@
-﻿using System.Net;
-
-namespace MyWebApp.Infrastructure
+﻿namespace MyWebApp.Infrastructure
 {
-    public class AdminSafeListMiddleware
+    public class AdminSafeListMiddleware(
+        RequestDelegate next,
+        ILogger<AdminSafeListMiddleware> logger,
+        IpSafeList safeList)
     {
-        private readonly RequestDelegate _next;
-        private readonly ILogger<AdminSafeListMiddleware> _logger;
-        private readonly List<IPAddress> _ipAddresses;
-        private readonly List<IPNetwork> _ipNetworks;
-
-        public AdminSafeListMiddleware(RequestDelegate next, ILogger<AdminSafeListMiddleware> logger, IpSafeList safeList)
-        {
-            _ipAddresses = safeList.IpAddresses.Split(';').Select(IPAddress.Parse).ToList();
-            _ipNetworks = safeList.IpNetworks.Split(';').Select(IPNetwork.Parse).ToList();
-
-            _next = next;
-            _logger = logger;
-        }
-
-        public async Task Invoke(HttpContext context)
+        public Task Invoke(HttpContext context)
         {
             var remoteIp = context.Connection.RemoteIpAddress;
             if (remoteIp == null)
             {
                 throw new ArgumentException("Remote IP is NULL, may due to missing ForwardedHeaders.");
             }
-            _logger.LogDebug("Remote IpAddress: {RemoteIp}", remoteIp);
+            logger.LogDebug("Remote IpAddress: {RemoteIp}", remoteIp);
 
             if (remoteIp.IsIPv4MappedToIPv6)
             {
                 remoteIp = remoteIp.MapToIPv4();
             }
 
-            if (!_ipAddresses.Contains(remoteIp) && !_ipNetworks.Any(x => x.Contains(remoteIp)))
+            if (!safeList.IsSafeIp(remoteIp))
             {
-                _logger.LogWarning("Forbidden Request from IP: {remoteIp}", remoteIp);
+                logger.LogWarning("Forbidden Request from IP: {remoteIp}", remoteIp);
                 context.Response.StatusCode = StatusCodes.Status403Forbidden;
-                return;
+                return Task.CompletedTask;
             }
 
-            await _next.Invoke(context);
+            return next.Invoke(context);
         }
     }
 }
